+++ commands/src/main/java/net/apartium/cocoabeans/commands/lexer/CommandTokenType.java
@@ -1,0 +1,9 @@
+package net.apartium.cocoabeans.commands.lexer;
+
+import org.jetbrains.annotations.ApiStatus;
+
+@ApiStatus.AvailableSince("0.0.37")
+public enum CommandTokenType {
+    KEYWORD,
+    ARGUMENT_PARSER
+}